moved the detect version type function to parser utils
